fixing deployment issue and inq page
