+++ src/app/inquiry/page.js
@@ -1,6 +1,16 @@
-import AdminReport from "@/pages/InquiryPage";
-import AdminTable from "@/components/InquiryTable";
+import InquiryTable from "@/components/InquiryTable";
+import Navbar from "@/components/Navbar";
 
 export default function Page() {
-  return (<AdminReport />);
+  const reports = [
+    { business: "Bob Ryan", service: "Personal Training", date: "2025-03-10", reviewLink: "/review/1" },
+    { business: "Sinead O'Boyle", service: "Yoga Class", date: "2025-03-08", reviewLink: "/review/2" },
+    { business: "Hannah Montanna", service: "Smoothie Making", date: "2025-03-09", reviewLink: "/review/2" },
+    { business: "Ryan Renaults", service: "Sports Massage", date: "2025-04-14", reviewLink: "/review/2" }
+  ];
+
+  return (<div>
+    <Navbar />
+    <InquiryTable reports={reports} />
+  </div>);
 }
