+++ src/styles/InquiryTable.css
@@ -2,7 +2,7 @@
 .heading {
   text-align: center;
   font-weight: bold;
-  margin: 20px 0;
+  margin-top: 120px;
   color: #ff5125;
 }
